+++ packages/lib/modules/vebal/vote/VoteList/VoteListLayout.tsx
@@ -12,11 +12,11 @@ import {
   useFilterTagsVisible,
   VoteListFilters,
 } from '@repo/lib/modules/vebal/vote/VoteList/VoteListFilters'
-import { poolTypeLabel } from '@repo/lib/modules/pool/PoolList/usePoolListQueryState'
 import { SelectedPoolsMenu } from '@repo/lib/modules/vebal/vote/VoteList/SelectedPoolsMenu'
 import { useVotes } from '@repo/lib/modules/vebal/vote/Votes/VotesProvider'
 import { StaticToast } from '@repo/lib/shared/components/toasts/StaticToast'
 import { useMemo } from 'react'
+import { poolTypeLabel } from '@repo/lib/modules/pool/pool.helpers'
 
 export function VoteListLayout() {
   const {
